Forcing interface reset if device exists
When a profile is being brought up, but the interface already
exists, netctl exits to not disturb any existing configuration.
When DHCP fails on a VLAN interface, the interface persists even
though the profile is not started successfully, and subsequent
attempts at starting or resetting the profile fail.

This patch does two things:
 a) centralize `ip link delete' in bring_interface_down
 b) attempt once to bring_interface_down if it already exists
    and try again

Signed-off-by: Dennis Herbrich <dennis.herbrich@hytera.de>